Moved several text methods to a utility class
